mdcode: rename semantic KC files to knowledge_catalog
The emitter/publisher pair was catalog.ts + kc.ts, which both read as
'Knowledge Catalog' and collide with the CatalogClient/CatalogManifest
family. Rename by product, mirroring the BigQuery pair:

  bigquery.ts / deploy.ts   (BigQuery: emit / publish)
  knowledge_catalog.ts / deploy_knowledge_catalog.ts   (KC: emit / publish)

Tests and goldens follow the same scheme (*.knowledge_catalog.golden.json,
parallel to *.bigquery.golden.sql). Pure rename plus import/comment
updates; no behavior change.
