+++ toolbox/mdcode/tests/libts/semantic/fixtures/sales_bq_graph_target.knowledge_catalog.golden.json
